Add simple workflow graph to run pages.
