+++ apps/web/app/console/product/run/[productRunId]/client.tsx
@@ -30,6 +30,10 @@ import {
 } from '../../_hooks'
 import { canManageConsoleChildResource } from '../../../../../utils/access-control'
 import { WorkflowDagChart } from '../../../../../components/workflow-dag-chart'
+import {
+  SimpleWorkflowDagChart,
+  type WorkflowDagSimple,
+} from '../../../../../components/simple-workflow-dag-chart'
 
 const ProductRunDetails = () => {
   const productRunQuery = useProductRun()
@@ -157,6 +161,13 @@ const ProductRunDetails = () => {
               workflowDag={productRun?.workflowDag}
               runType="product"
             />
+            {productRun?.workflowDagSimple && (
+              <SimpleWorkflowDagChart
+                workflowDagSimple={
+                  productRun.workflowDagSimple as WorkflowDagSimple
+                }
+              />
+            )}
           </CrudForm>
         )}
       </div>
